backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Pillow==9.0.0
